worker/requirements.txt
# Worker核心依赖
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
peft>=0.6.0
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.0.0
sentencepiece
protobuf

# CLI工具
rich>=13.0.0

# 直连服务器
fastapi>=0.100.0
uvicorn>=0.23.0
aiohttp>=3.9.0

# 高性能推理后端（可选，按需安装）
# sglang>=0.1.0          # SGLang 推理后端
# vllm>=0.4.0            # vLLM 推理后端

# 分布式通信
grpcio>=1.60.0
grpcio-tools>=1.60.0

# 缓存和序列化
redis>=5.0.0
lz4>=4.3.0              # 快速压缩

# 可选: CUDA量化支持 (需要NVIDIA GPU)
# bitsandbytes>=0.41.0

